+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Ethernet protocol parser.
 - Fixed VLAN Tag attribute setters.
 - Added IPv4 protocol parser.
+- Added ICMP protocol parser.